+++ src/shaft.h
@@ -40,13 +40,13 @@ public:
 	csArray<int> elevators; //if elevator shaft, this is the list of elevators that use this shaft
 	csArray<int> stairs; //if stairwell, this lists the stairs that use this shaft
 	bool InsideShaft; //true if user/camera is in the shaft
-	double bottom; //shaft base
-	double top; //shaft top
+	float bottom; //shaft base
+	float top; //shaft top
 
-	Shaft(int number, int type, double CenterX, double CenterZ, int _startfloor, int _endfloor);
+	Shaft(int number, int type, float CenterX, float CenterZ, int _startfloor, int _endfloor);
 	~Shaft();
-	int AddWall(int floor, const char *name, const char *texture, double thickness, double x1, double z1, double x2, double z2, double height1, double height2, double voffset1, double voffset2, double tw, double th);
-	int AddFloor(int floor, const char *name, const char *texture, double thickness, double x1, double z1, double x2, double z2, double voffset1, double voffset2, double tw, double th);
+	int AddWall(int floor, const char *name, const char *texture, float thickness, float x1, float z1, float x2, float z2, float height1, float height2, float voffset1, float voffset2, float tw, float th);
+	int AddFloor(int floor, const char *name, const char *texture, float thickness, float x1, float z1, float x2, float z2, float voffset1, float voffset2, float tw, float th);
 	void Enabled(int floor, bool value);
 	bool IsShaft(csRef<iMeshWrapper> test);
 	void EnableWholeShaft(bool value);
